LayoutTests/imported/w3c/web-platform-tests/IndexedDB/fire-success-event-exception-expected.txt

CONSOLE MESSAGE: line 28: Error
CONSOLE MESSAGE: line 37: Error
CONSOLE MESSAGE: line 49: Error
CONSOLE MESSAGE: line 59: Error

PASS Exception in success event handler on request 
PASS Exception in success event listener on request 
PASS Exception in second success event listener on request 
FAIL Exception in first success event listener, tx active in second assert_equals: Active check should either not throw anything, or throw TransactionInactiveError expected "TransactionInactiveError" but got "InvalidStateError"
